requirements.txt
pandas
numpy
yfinance
streamlit
sqlalchemy
matplotlib
seaborn
openpyxl
pyyaml
tomli
toml
requests